Add Enemy base class
Close#55
